# db/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Contacts API",
    "description": "Contacts API developed by memx86",
    "contact": {
      "email": "memx86@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.2"
  },
  "servers": [
    {
      "url": "https://memx86-contacts.herokuapp.com/api"
    }
  ],
...
  "paths": {
...
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "log in",
        "description": "Log in user\n",
        "operationId": "loginUser",
        "requestBody": {
          "description": "user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "400": {
            "description": "Validation error | Email or password is wrong | Please verify your email"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/verify/{verificationToken}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "verify user",
        "description": "Provide verification token to get activation link via email\n",
        "operationId": "verifyUser",
        "parameters": [
          {
            "name": "verificationToken",
            "in": "path",
            "description": "user's verification token",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Verification successful"
          },
          "400": {
            "description": "Verification has already been passed"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/verify": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "reverify user",
        "description": "Provide  email to resend verification link\n",
        "operationId": "reVerifyUser",
        "requestBody": {
          "description": "user's email to reverify",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Verify"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification email sent"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/users/logout": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "log user out",
        "description": "Logout user\n",
        "operationId": "logoutUser",
        "responses": {
          "204": {
            "description": "successfull logout"
          },
          "400": {
            "description": "Wrong token type"
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@mail.com"
          },
          "subscription": {
            "type": "string",
            "example": "starter"
          },
          "avatarURL": {
            "type": "string",
            "example": "https://s.gravatar.com/avatar/e1102c33ff44561b951fc9fde8350062"
          }
        }
      },
...
      "LogIn": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@mail.com"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
      "Verify": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "john.doe@mail.com"
          }
        }
      },
...
      "inline_response_200": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "JWT token"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}